Attempt to remove some unnecessary asio includes
